DSS_Code/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.20.0
transformers>=4.30.0
scikit-learn>=1.2.0
numpy>=1.24.0
matplotlib>=3.7.0
umap-learn>=0.5.3
Pillow>=9.5.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0